data_engineering/03_Kafka/producer&consumer/requirements.txt
kafka-python==2.0.4
confluent-kafka==2.3.0